requirements.txt
langchain 
llama-index 
networkx 
pyvis 
transformers 
torch 
faiss-cpu